# LibJS/JIT: Add fastpath for set variable