This one was hard to benchmark for me again, since the benchmarks didn't have meaningful differences (i might need to lock the CPU frequency or something). So here is a synthetic benchmark:
```js
for(let n = 0.0; n < 1000000000; n+=2) {
}
```
PR:
```
Benchmark 1: js ./test.js
  Time (mean ± σ):      3.422 s ±  0.019 s    [User: 3.411 s, System: 0.002 s]
  Range (min … max):    3.401 s …  3.466 s    10 runs
```
Master:
```
Benchmark 1: js ./test.js
  Time (mean ± σ):     21.617 s ±  0.277 s    [User: 21.538 s, System: 0.007 s]
  Range (min … max):   21.305 s … 22.208 s    10 runs
  ```